Lab2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-slugify==8.0.4
